packages/ai-core/requirements.txt
fastapi
uvicorn[standard]
pandas
scikit-learn
pydantic

pre-commit
ruff
black
isort
nbstripout
pytest

sentencepiece
sacremoses
faiss-cpu
pyyaml
tqdm

psycopg2-binary>=2.9
python-dotenv>=1.0
transformers>=4.44
torch>=2.2
numpy>=1.26
googletrans==4.0.0-rc1
psycopg[binary]>=3.1